// src/gen/test/v1/test_pb.ts
// @generated by protoc-gen-es v2.13.0 with parameter "target=ts"
// @generated from file test/v1/test.proto (package test.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file test/v1/test.proto.
 */
export const file_test_v1_test: GenFile = /*@__PURE__*/
  fileDesc("ChJ0ZXN0L3YxL3Rlc3QucHJvdG8SB3Rlc3QudjEiGwoLQ2FsbFJlcXVlc3QSDAoEZGF0YRgBIAEoCSIcCgxDYWxsUmVzcG9uc2USDAoEZGF0YRgBIAEoCTJHCgtUZXN0U2VydmljZRI4CgRDYWxsEhQudGVzdC52MS5DYWxsUmVxdWVzdBoVLnRlc3QudjEuQ2FsbFJlc3BvbnNlIgOQAgFiBnByb3RvMw");

/**
 * @generated from message test.v1.CallRequest
 */
export type CallRequest = Message<"test.v1.CallRequest"> & {
  /**
   * @generated from field: string data = 1;
   */
  data: string;
};

/**
 * Describes the message test.v1.CallRequest.
 * Use `create(CallRequestSchema)` to create a new message.
 */
export const CallRequestSchema: GenMessage<CallRequest> = /*@__PURE__*/
  messageDesc(file_test_v1_test, 0);

/**
 * @generated from message test.v1.CallResponse
 */
export type CallResponse = Message<"test.v1.CallResponse"> & {
  /**
   * @generated from field: string data = 1;
   */
  data: string;
};

/**
 * Describes the message test.v1.CallResponse.
 * Use `create(CallResponseSchema)` to create a new message.
 */
export const CallResponseSchema: GenMessage<CallResponse> = /*@__PURE__*/
  messageDesc(file_test_v1_test, 1);

/**
 * @generated from service test.v1.TestService
 */
export const TestService: GenService<{
  /**
   * @generated from rpc test.v1.TestService.Call
   */
  call: {
    methodKind: "unary";
    input: typeof CallRequestSchema;
    output: typeof CallResponseSchema;
  },
}> = /*@__PURE__*/
  serviceDesc(file_test_v1_test, 0);